requirements.txt
asgiref==3.4.1
cssselect2==0.4.1
Django==3.2.13
django-ckeditor==6.3.2
django-js-asset==2.0.0
djangorestframework==3.13.1
fpdf==1.7.2
jsonfield==3.1.0
lxml==4.9.0
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.8
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
tinycss2==1.1.1
typing_extensions==4.1.1
urllib3==1.26.9
webencodings==0.5.1
